requirements.txt
# Core dependencies for Iteration 1 workflow implementation
openai-agents>=0.4.2
python-dotenv>=1.0.0
pydantic>=2.7.0
httpx>=0.27.0
